% can be used in conjunction with arithmetic operations
the % symbol to still work as an equals operand, but if there is an arithmetic operator in the number bar then it converts the number right of the operand into a percent of the number to the left of the operand.  afterwards do not turn them into a percent, and allow more operations to be done.